Add Show Input Sketch Point to Experimental panel; Draw directly in Canvas rather than Workspace to access setting
