+++ libs/vgc/workspace/edge.cpp
@@ -327,92 +327,6 @@ enum class ShowEdgeDirection {
     AllEdges
 };
 
-// Note: we do not bother to implement any caching for this since it is
-// mostly for debugging purposes and performance is not critical.
-//
-void doPaintInputSketchPoints(
-    graphics::Engine* engine,
-    const VacKeyEdge& edge,
-    graphics::GeometryViewPtr& geometry) {
-
-    using geometry::Mat3d;
-    using geometry::Vec2d;
-    using geometry::Vec2dArray;
-    using geometry::Vec2f;
-    namespace ds = dom::strings;
-
-    // Get the positions of the input sketch points, in widget coordinates (at
-    // time of sketch)
-    //
-    dom::Element* e = edge.domElement();
-    const auto& positions = e->getAttribute(ds::inputpositions).get<Vec2dArray>();
-    Int n = positions.length();
-    if (n <= 0) {
-        return;
-    }
-
-    // Get the widget to scene transform matrix
-    //
-    const auto& transform = e->getAttribute(ds::inputtransform).get<Mat3d>();
-
-    // Create the graphics resource
-    //
-    if (!geometry) {
-        geometry = engine->createDynamicTriangleStripView(
-            graphics::BuiltinGeometryLayout::XYDxDy_iXYRotWRGBA);
-    }
-
-    // Compute, in scene coordinates, the corners of a square centered at the
-    // origin, scaled and rotated such that it has the same size and
-    // orientation as a pixel when the edge was first sketched. The "disp"
-    // component is used to be able to apply a small screen-space displacement,
-    // so that we can paint a thin border of w pixels around the square.
-    //
-    //  x-----------x
-    //  | x-------x |
-    //  | |       | |
-    //  | |       |w|
-    //  | |       | |
-    //  | x-------x |  <- cornerPos
-    //  x-----------x  <- cornerPos + cornerDisp * w
-    //
-    constexpr float sqrt2 = 1.4142135f;
-    struct PosAndDisp {
-        PosAndDisp(const Vec2d& pos)
-            : pos_(pos)
-            , disp_(sqrt2 * pos_.normalized()) {
-        }
-        Vec2f pos_;
-        Vec2f disp_;
-    };
-    core::Array<PosAndDisp> sharedInstData = {
-        PosAndDisp(transform.transformLinear({-0.5, -0.5})),
-        PosAndDisp(transform.transformLinear({0.5, -0.5})),
-        PosAndDisp(transform.transformLinear({-0.5, 0.5})),
-        PosAndDisp(transform.transformLinear({0.5, 0.5}))};
-
-    // We draw two quads for each input sketch point:
-    // - one with a small screen-space displacement w
-    // - one without screen-space displacement (w = 0)
-    //
-    const core::Color& c = colors::selection;
-    core::FloatArray perInstData;
-    for (Int i = 0; i < n; ++i) {
-        float w = 1.f;
-        Vec2d pWidget = positions[i];
-        Vec2f pScene = Vec2f(transform.transformAffine(pWidget));
-        perInstData.extend({pScene.x(), pScene.y(), 1.f, w, c.r(), c.g(), c.b(), 1.f});
-        perInstData.extend({pScene.x(), pScene.y(), 1.f, 0.f, 1.f, 1.f, 1.f, 1.f});
-        //                     X           Y        Rot   W    R    G    B    A
-    }
-
-    engine->updateBufferData(geometry->vertexBuffer(0), std::move(sharedInstData));
-    engine->updateBufferData(geometry->vertexBuffer(1), std::move(perInstData));
-
-    engine->setProgram(graphics::BuiltinProgram::ScreenSpaceDisplacement);
-    engine->drawInstanced(geometry);
-}
-
 } // namespace
 
 void VacKeyEdge::onPaintDraw(
@@ -450,8 +364,6 @@ void VacKeyEdge::onPaintDraw(
     bool shouldPaintEdgeDirection =
         (showEdgeDirection == ShowEdgeDirection::AllEdges)
         || (isSelected && showEdgeDirection == ShowEdgeDirection::SelectedEdges);
-    bool shouldPaintInputSketchPoints = isSelected;
-    //                           TODO: `&& options::showInputSketchPoints();`
 
     bool needsCenterlineGeometry =
         shouldPaintOutline || shouldPaintOffsetLine0 || shouldPaintOffsetLine1;
@@ -738,9 +650,6 @@ void VacKeyEdge::onPaintDraw(
     if (isSelected) {
         engine->setProgram(graphics::BuiltinProgram::ScreenSpaceDisplacement);
         engine->draw(graphics.selectionGeometry());
-        if (shouldPaintInputSketchPoints) {
-            doPaintInputSketchPoints(engine, *this, inputSketchPointsGeometry_);
-        }
     }
     else if (shouldPaintOutline) {
         engine->setProgram(graphics::BuiltinProgram::ScreenSpaceDisplacement);
